Refactor role management by renaming Admin folder to RBAC for better naming
