+++ src/rbac/rbac.module.ts
@@ -2,8 +2,8 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from '../users/entities/user.entity';
 import { Role } from '../roles/entities/role.entity';
-import { AdminService } from './admin.service';
-import { AdminController } from './admin.controller';
+import { AdminService } from './rbac.service';
+import { AdminController } from './rbac.controller';
 import { PermissionsGuard } from './guards/permissions.guard';
 import { AuthModule } from '../auth/auth.module';
 import { UsersModule } from '../users/users.module';
